pyproject.toml
[tool.poetry]
name = "ga4-etl"
version = "0.1.0"
description = "GA4データをBigQueryから抽出し、解析しやすいデータ構造に変換して再度BigQueryにロードするETLパイプライン"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-bigquery = "3.11.4"
google-cloud-storage = "2.13.0"
pandas = "2.0.3"
pandas-gbq = "0.19.2"
pyarrow = "14.0.1"
google-cloud-run = "0.10.5"
slack-sdk = "3.24.0"
python-dotenv = "1.0.0"
pyyaml = "6.0.1"
ruff = "^0.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
extend-select = ["I"]

[tool.poetry.scripts]
ga4-etl = "main:main"
